Store positions in int64 matrix. This to support files over 4.2 GB

+++ tpx3format/read.py
@@ -17,9 +17,9 @@ def read_raw(file_name, cores):
     f = open(file_name, "rb")
     estimate = os.fstat(f.fileno()).st_size / 8
 
-    # Allocate an array to hold positions of packages
+    # Allocate an array to hold positions of packages. Using int64 to support files over 4.2 GB
     max_positions = 100
-    positions = np.empty((max_positions, 3), dtype='uint32')
+    positions = np.empty((max_positions, 3), dtype='int64')
 
     # Check if we have a loadable ToT correction file
     check_tot_correction(lib.config.settings.hits_tot_correct_file)
